Include HTML modules, and refactor Ingredients + date utils into modules
