+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-files: ^js/app.js|css/styles.css|index.html|.github/|.pre-commit-config.yaml # only website specific files
+files: ^js/app.js|js/modules/|js/components/|css/styles.css|index.html|.github/|.pre-commit-config.yaml # only website specific files
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v3.4.0
