+++ index.html
@@ -37,7 +37,7 @@
 <footer class="notification is-warning has-text-centered mb-0">
     <b>Important:</b> This page shows the menus of the current week in a human-readable format. If you are searching for how to use the API, please take a look at the <a href="https://github.com/TUM-Dev/eat-api">README</a>.
 </footer>
-<script src="js/app.js"></script>
+<script src="js/app.js" type="module"></script>
 </body>
 
 </html>
